+++ TMS.Api/TMS.Api/Models/Order.cs
@@ -5,7 +5,7 @@ namespace TMS.Api.Models;
 
 public partial class Order
 {
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
 
     public decimal TotalPrice { get; set; }
 
@@ -13,9 +13,9 @@ public partial class Order
 
     public DateTime OrderedAt { get; set; }
 
-    public int TicketCategoryId { get; set; }
+    public long TicketCategoryId { get; set; }
 
-    public int CustomerId { get; set; }
+    public long CustomerId { get; set; }
 
     public virtual User Customer { get; set; } = null!;
